Committing vehicles_controller.rb, welcome.html.erb, routes.rb e delete.html.erb

+++ app/controllers/vehicles_controller.rb
@@ -21,27 +21,28 @@ class VehiclesController < ApplicationController
 
     def show
         @vehicle = Vehicle.find(params[:id])
-
+    end
+    
+    def destroy
+        @vehicle = Vehicle.find_by(placa: params[:placa])
+      
+        if @vehicle
+          if @vehicle.destroy
+            flash[:notice] = 'Registro de veículo excluído com sucesso.'
+            redirect_to delete_vehicle_path
+          else
+            flash[:error] = 'Erro ao excluir o registro de veículo.'
+            redirect_to delete_vehicle_path(placa: @vehicle.placa)
+          end
+        else
+          flash[:error] = 'Veículo não encontrado.'
+          redirect_to delete_vehicle_path
+        end
+      end
+      
     private
 
     def vehicle_params
         params.require(:vehicle).permit(:placa, :nome_veiculo, :dono_do_veiculo, :hora_entrada, :hora_saida, :total_a_pagar_por_veiculo, :subtotal)
     end
-
-    def delete
-        placa = params[:placa]
-        @vehicle = Vehicle.find_by(placa: placa)
-    end
-    
-    def destroy
-        @vehicle = Vehicle.find(params[:id])
-        if @vehicle.destroy
-            flash[:notice] = 'Registro de veículo excluido com sucesso.'
-            redirect_to delete_vehicle_path
-        else
-            flash[:error] = 'Erro ao excluir o registro de veículo.'
-            redirect_to delete_vehicle_path(placa: @vehicle.placa)
-        end
-    end   
-
 end
